pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv_sql_analyst"
version = "0.1.0"
description = "OpenEnv SQL Data Analyst Agent"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "openai"
]

[project.scripts]
server = "server.app:main"